ml-api/requirements.txt
flask==3.0.3
pillow==10.3.0
numpy==1.26.4
tflite-runtime==2.14.0
gunicorn==22.0.0